Authorization

+++ backend/routes/posts.js
@@ -33,9 +33,9 @@ router.post('',checkAuth,multer({storage:storage}).single("image"),(req,res,next
         title: req.body.title,
         content:req.body.content,
         imagePath:url+"/images/"+req.file.filename,
-        creator : req.userData.id
+        creator:req.userData.userId
     });
-    
+  
     post.save().then(responseData=>{
         console.log(responseData);
         res.status(201).json({
@@ -64,17 +64,26 @@ router.put('/:id',checkAuth,multer({storage:storage}).single("image"),(req,res,n
         _id:req.params.id,
         title : req.body.title,
         content: req.body.content,
-        imagePath:imagePath
+        imagePath:imagePath,
+        creator:req.userData.userId
     })
 
     console.log('Post',post.imagePath);
-    Post.updateOne({_id:req.params.id},post)
+    Post.updateOne({_id:req.params.id,creator:req.userData.userId},post)
     .then((result)=>{
         console.log('Result',result)
-        res.status(200).json({
-            message:"update Successfully",
-            imagePath:post.imagePath
-        })
+        if(result.nModified > 0){
+            res.status(200).json({
+                message:"update Successfully",
+                imagePath:post.imagePath
+            })
+        }
+        else {
+            res.status(401).json({
+                message:"Not Authorized"
+            })
+        }
+        
     })
 })
 
@@ -115,12 +124,19 @@ router.get('',(req,res,next)=>{
 
 
 router.delete('/:id',checkAuth,(req,res,next) =>{
- Post.deleteOne({_id:req.params.id})
+ Post.deleteOne({_id:req.params.id,creator:req.userData.userId})
  .then((result)=>{
   console.log(result)
-  res.status(200).json({
-    message :'Post Deleted'
-})
+  if(result.n > 0){
+    res.status(200).json({
+        message:"delete Successfully"
+    })
+}
+else {
+    res.status(401).json({
+        message:"Not Authorized"
+    })
+}
  })
  
 })
